Chore: improve error handling
Chore: remove and scope comments and further cleanup
