+++ client/src/App.vue
@@ -35,8 +35,7 @@ export default {
       return Promise.resolve()
         .then(() => {
           if (!(apiUrl || sessionStorage.getItem(sKeys.API_URL))) {
-            console.log('apiURl >> ', apiUrl)
-            throw 'Parameters error'
+            return Promise.reject('Please provide the API\'s url')
           }
         })
         .then(() => {
@@ -44,19 +43,17 @@ export default {
             return StoreHelper.fetchToken(refreshToken, apiUrl)
               .then(() => {
                 const token = StoreHelper.getAuthToken()
-                console.log('token >> ', token)
                 if (!token) {
-                  throw 'Refresh token is expired'
+                  return Promise.reject('Refresh token is expired')
                 } else {
                   return StoreHelper.fetchTokenData(token, apiUrl)
                 }
               })
           } else if (authToken) {
-            console.log('hasAuthToken >> ', authToken)
             setAuthHeader()
             return StoreHelper.fetchTokenData(authToken, apiUrl)
           } else {
-            throw 'Parameters Error'
+            return Promise.reject('Parameters Error')
           }
         })
         .then(() => {
@@ -64,7 +61,7 @@ export default {
         })
         .then((payload) => {
           if (!payload) {
-            throw 'Error when fetching token data'
+            return Promise.reject('Error when fetching token data')
           } else {
             visitId = payload.visitId
           }
