+++ server/src/server/api.js
@@ -16,6 +16,7 @@ import UserController from '../mcr/user/user-controller.js'
 import VisitController from '../mcr/visit/visit-controller'
 import SeedDataController from '../mcr/seed/seedData-controller'
 import AuthController from '../mcr/auth/auth-controller'
+import { validatorMiddleware } from '../helpers/middleware'
 
 // Sessions and session cookies
 // express-session stores session data here on the server and only puts session id in the cookie
@@ -24,29 +25,6 @@ const FileStore = require('session-file-store')(session)
 const uuid = require('uuid/v4')
 
 const debug = require('debug')('server')
-
-const validatorMiddleware = (req, res, next) => {
-  const authController = new AuthController()
-  if (req && req.headers.authorization) {
-    try {
-      const result = authController.authenticate(req.headers.authorization)
-      const { visitId } = result
-      console.log('result >> ', result)
-      if (visitId) {
-        console.log('passingNext!!!')
-        next()
-      } else {
-        res.status(401).send('Invalid token!')
-      }
-    } catch (err) {
-      console.log('validatorMiddleware caught ', err)
-      res.status(401).send(err)
-    }
-  } else {
-    console.log('validatorMiddleware else ', req)
-    res.status(401).send('A token is required')
-  }
-}
 
 export function apiMiddle (app, config) {
   const fileStoreOptions = {}
